requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
transformers==4.57.1
sentence-transformers==5.1.2
faiss-cpu==1.13.0
numpy==2.3.5
pydantic==2.12.4
